Renamed package of the IndQNER dataset classes.
